--- a/cmd/flow-dps-live/main.go
+++ b/cmd/flow-dps-live/main.go
@@ -323,7 +323,7 @@ func run() int {
 			return failure
 		}
 		defer file.Close()
-		load = loader.FromCheckpoint(file)
+		load = loader.FromCheckpoint(log, file)
 	} else if flagCheckpoint != "" {
 		file, err := os.Open(flagCheckpoint)
 		if err != nil {
@@ -331,7 +331,7 @@ func run() int {
 			return failure
 		}
 		defer file.Close()
-		initialize := loader.FromCheckpoint(file)
+		initialize := loader.FromCheckpoint(log, file)
 		load = loader.FromIndex(log, storage, indexDB,
 			loader.WithInitializer(initialize),
 			loader.WithExclude(loader.ExcludeAtOrBelow(first)),